new local database setted

+++ src/main/java/com/example/cardapio/food/FoodRepository.java
@@ -2,6 +2,6 @@ package com.example.cardapio.food;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoodRepository extends JpaRepository<Food, Integer> {
+public interface FoodRepository extends JpaRepository<Food, Long> {
 
 }
