+++ src/main/java/com/example/cardapio/food/FoodResponseDTO.java
@@ -1,6 +1,6 @@
 package com.example.cardapio.food;
 
-public record FoodResponseDTO(Integer id, String tile, String image, Integer price) {
+public record FoodResponseDTO(Long id, String tile, String image, Integer price) {
     public FoodResponseDTO(Food food){
         this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
     }
